requirements.txt
# SCA - Python packages with known critical/high vulnerabilities
Django==2.0.0
Flask==0.11.1
requests==2.6.0
urllib3==1.21.1
Pillow==5.0.0
cryptography==1.9
PyYAML==3.12
lxml==3.7.0
Jinja2==2.8.0
Werkzeug==0.11.0
sqlalchemy==1.1.0
celery==3.1.18
paramiko==1.16.0
Twisted==16.0.0
pyOpenSSL==16.0.0
boto3==1.4.0
numpy==1.13.0
scipy==0.19.0
aiohttp==2.0.0
httplib2==0.9.1
pygments==2.1.3
reportlab==3.3.0
nltk==3.2.1
setuptools==20.0.0
pip==9.0.0
